backend/requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy
fastapi.middleware.cors
numpy
BytesIO
Image
Nodejs
Reactjs
Material-Ui
web-vitals
axios
